requirements.txt
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
requests==2.31.0
